adaptive-learning-platform/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sentence-transformers>=2.2.0
chromadb>=0.4.0
transformers>=4.36.0
torch>=2.2.0
safetensors>=0.4.0
accelerate>=0.25.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
redis>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jinja2==3.1.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
